修改命令发送接口，徐州命令只传给LC 增加url参数缺失异常处理 增加rmi节点实体类获取id的方法 将自定义类型全部去掉

+++ afc-monitor-api/src/main/java/com/insigma/afc/monitor/controller/NodeStatusController.java
@@ -93,7 +93,7 @@ public class NodeStatusController {
             equStatus.setStatus(DeviceStatus.getInstance().getNameByValue(status) + "/" + status);
             equStatus.setOnline(equStatusViewItem.getOnline());
             equStatus.setUpdateTime(DateTimeUtil.formatDate(equStatusViewItem.getUpdateTime()));
-            equStatus.setName(topologyService.getNodeText(equStatusViewItem.getNodeId().longValue()).getData());
+            equStatus.setName(topologyService.getNodeText(equStatusViewItem.getNodeId()).getData());
             equStatusList.add(equStatus);
         }
         return Result.success(equStatusList);
